pyproject.toml
[project]
name = "python-agentframework-demos"
version = "0.1.0"
description = "Collection of Python examples for Microsoft Agent Framework"
requires-python = ">=3.10"
dependencies = [
    "azure-identity",
    "openai>=1.109.1",
    "python-dotenv",
    "pydantic",
    "rich",
    "dotenv-azd",
    "aiohttp",
    "faker",
    "fastmcp",
    "opentelemetry-exporter-otlp-proto-grpc",
    "azure-monitor-opentelemetry",
    "psycopg[binary]",
    "pgvector",
    "azure-ai-evaluation[redteam]>=1.15.0",
    "agent-framework-core @ git+https://github.com/microsoft/agent-framework.git@fc9c81b0b11170bdab8fa2d42bb96981e65fd270#subdirectory=python/packages/core",
    "agent-framework-devui @ git+https://github.com/microsoft/agent-framework.git@fc9c81b0b11170bdab8fa2d42bb96981e65fd270#subdirectory=python/packages/devui",
    "agent-framework-redis @ git+https://github.com/microsoft/agent-framework.git@fc9c81b0b11170bdab8fa2d42bb96981e65fd270#subdirectory=python/packages/redis",
    "agent-framework-mem0 @ git+https://github.com/microsoft/agent-framework.git@fc9c81b0b11170bdab8fa2d42bb96981e65fd270#subdirectory=python/packages/mem0",
    "agent-framework-azure-ai-search @ git+https://github.com/microsoft/agent-framework.git@fc9c81b0b11170bdab8fa2d42bb96981e65fd270#subdirectory=python/packages/azure-ai-search",
]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
